NFC: move ast passes to ast/passes dir
As the amount of passes grows a bit of structure becomes useful.
