+++ src/ast/CMakeLists.txt
@@ -1,18 +1,19 @@
 add_library(ast
   ast.cpp
   attachpoint_parser.cpp
-  codegen_llvm.cpp
   fake_map.cpp
-  field_analyser.cpp
   int_parser.cpp
   irbuilderbpf.cpp
   pass_manager.cpp
-  portability_analyser.cpp
-  printer.cpp
-  resource_analyser.cpp
-  semantic_analyser.cpp
   signal.cpp
   visitors.cpp
+
+  passes/field_analyser.cpp
+  passes/portability_analyser.cpp
+  passes/printer.cpp
+  passes/resource_analyser.cpp
+  passes/semantic_analyser.cpp
+  passes/codegen_llvm.cpp
 )
 
 target_include_directories(ast PUBLIC ${CMAKE_SOURCE_DIR}/src)
