+++ src/ast/passes/node_counter.h
