+++ src/ast/passes/resource_analyser.cpp
