+++ src/ast/passes/resource_analyser.h
