+++ src/fuzz_main.cpp
@@ -17,10 +17,10 @@
 #include <unistd.h>
 
 #include "ast/bpforc/bpforc.h"
-#include "ast/callback_visitor.h"
 #include "ast/clang_parser.h"
-#include "ast/field_analyser.h"
-#include "ast/semantic_analyser.h"
+#include "ast/passes/callback_visitor.h"
+#include "ast/passes/field_analyser.h"
+#include "ast/passes/semantic_analyser.h"
 #include "bpftrace.h"
 #include "codegen_llvm.h"
 #include "driver.h"
