+++ tests/codegen/common.h
@@ -7,12 +7,14 @@
 
 #include "../mocks.h"
 #include "ast/bpforc/bpforc.h"
-#include "ast/resource_analyser.h"
-#include "ast/semantic_analyser.h"
+
+#include "ast/passes/codegen_llvm.h"
+#include "ast/passes/resource_analyser.h"
+#include "ast/passes/semantic_analyser.h"
+
 #include "bpffeature.h"
 #include "bpftrace.h"
 #include "clang_parser.h"
-#include "codegen_llvm.h"
 #include "driver.h"
 #include "fake_map.h"
 #include "tracepoint_format_parser.h"
