+++ tests/codegen/regression.cpp
@@ -1,14 +1,6 @@
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
 
-#include "ast/bpforc/bpforc.h"
-#include "ast/resource_analyser.h"
-#include "ast/semantic_analyser.h"
-#include "bpftrace.h"
-#include "clang_parser.h"
-#include "codegen_llvm.h"
-#include "driver.h"
-
 #include "common.h"
 
 namespace bpftrace {
